lambda-cost-reporting-system/requirements.txt
# Core dependencies
boto3>=1.34.0
botocore>=1.34.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Email and templating
jinja2>=3.1.0

# Image processing for asset management
Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0
moto>=4.2.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# AWS Lambda Powertools for monitoring
aws-lambda-powertools>=2.25.0

# AWS CDK (for infrastructure)
aws-cdk-lib>=2.100.0
constructs>=10.3.0